adjust docstrings, rename html_reports_pipestat.py to reports.py
